Rename `peace_rt::commands` module to `peace_rt::cmds`.
