+++ crate/rt/src/cmds/states_current_display_cmd.rs
@@ -7,7 +7,7 @@ use peace_resources::{
 use peace_rt_model::{CmdContext, Error};
 use peace_rt_model_core::OutputWrite;
 
-use crate::StatesCurrentReadCmd;
+use crate::cmds::StatesCurrentReadCmd;
 
 /// Displays [`StatesCurrent`]s from storage.
 #[derive(Debug)]
