+++ crate/rt/src/lib.rs
@@ -6,9 +6,4 @@
 /// suitable number is.
 pub const BUFFERED_FUTURES_MAX: usize = 64;
 
-pub use crate::commands::{
-    DiffCmd, EnsureCmd, StatesCurrentDiscoverCmd, StatesCurrentDisplayCmd, StatesCurrentReadCmd,
-    StatesDesiredDiscoverCmd, StatesDesiredDisplayCmd, StatesDesiredReadCmd, StatesDiscoverCmd,
-};
-
-mod commands;
+pub mod cmds;
